Fix for issue 99
- The first mention of our product should be in full, ie IBM API
Connect. After that we can use the short name.
- It would be helpful to include more conceptual information here about
what rate limiting means.
- the description of Products/Plans could use a bit more information
- should change "1. Add a new rate-limiting plan to an existing API
product.", to "1. Create a new rate-limited Plan in an existing
Product."
- Remember that "Product" & "Plan" always have initial capitals (they
are specific terms in APIC).
- Avoid positional statements like "below" as they fail accessibility
checks
- Need to describe how to get to the Products page more explicitly, as
users could have clicked around on the site.
- Create a new rate limited plan section, Steps 5 & 6
- Change "catalog" to "Catalog"
- Stage & publish an updated product to the Sandbox catalog section,
Steps 1, 3, 5
- it would be useful to expand the title of this section to "Register a
new (consumer) application in the Developer Portal" to make it clear we
are moving to different tooling.
- needs some intro to the Dev Portal, updated line that describes the
"portal is ready" email.